Split valueNamesImpl into two pairs of files.
